Add rollup to build libruary. Add .editorconfig and .prettierrc. Update .npmignore. Move "./index.js" to "./src/index.js".
