+++ .npmignore
@@ -1,3 +1,6 @@
 .nyc_output
 coverage
 benchmark
+node_modules
+test
+rollup.config.js
