+++ benchmark/index.js
@@ -4,7 +4,7 @@ const deepCopy = require('deep-copy')
 const cloneDeep = require('lodash.clonedeep')
 const fastCopy = require('fast-copy').default
 const obj = require('./fixture.json')
-const clone = require('..')
+const clone = require('../src')
 const cloneDefaults = clone()
 const cloneProto = clone({proto: true})
 const cloneCircles = clone({circles: true})
